--- src/main/java/racinggame/Constant.java
@@ -1,11 +1,0 @@
-package racinggame;
-
-public class Constant {
-	static final int RANDOM_START_NUMBER = 0;
-	static final int RANDOM_END_NUMBER = 9;
-	static final int POSSIBLE_MOVE_NUMBER = 4;
-	static final String MOVE_EXPRESSION = "-";
-	static final String REG_EXP = "^[0-9a-zA-Z가-힣]{1,5}$";
-	static final String SPLIT_STR = ",";
-	static final String ADD_STR = ",";
-}
