+++ src/main/java/racinggame/cars/CarName.java
@@ -1,9 +1,11 @@
-package racinggame;
+package racinggame.cars;
 
-import static racinggame.Constant.*;
+import static racinggame.utils.Constant.*;
+import static racinggame.utils.RacingUtils.*;
 
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import racinggame.utils.Messages;
 
 public class CarName {
 	private String carName;
@@ -15,14 +17,9 @@ public class CarName {
 
 	private void validCarName(String carName) {
 		//영문, 한글, 숫자 포함 5자 이하 체크
-		if (!findMatches(Pattern.compile(REG_EXP), carName)) {
+		if (!findMatches(Pattern.compile(CAR_NAME_REG_EXP), carName)) {
 			throw new IllegalArgumentException(Messages.CAR_NAME_NOT_VALID.getValues());
 		}
-	}
-
-	private boolean findMatches(Pattern pattern, CharSequence input) {
-		Matcher m = pattern.matcher(input);
-		return m.find();
 	}
 
 	public String getCarName() {
