+++ src/main/java/racinggame/utils/Constant.java
@@ -1,0 +1,11 @@
+package racinggame.utils;
+
+public class Constant {
+	public static final int RANDOM_START_NUMBER = 0;
+	public static final int RANDOM_END_NUMBER = 9;
+	public static final int POSSIBLE_MOVE_NUMBER = 4;
+	public static final String MOVE_EXPRESSION = "-";
+	public static final String CAR_NAME_REG_EXP = "^[0-9a-zA-Z가-힣]{1,5}$";
+	public static final String SPLIT_STR = ",";
+	public static final String ADD_STR = ",";
+}
