+++ src/test/java/racinggame/CarNameTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import racinggame.cars.CarName;
+import racinggame.utils.Messages;
+
 class CarNameTest {
 
 	@Test
